split manager queue to composable
